Introduce a common parent class for NumericalColumn and DecimalColumn (#8278)
Adds a common parent class for NumericalColumn and DecimalColumn that implements the common operations (mainly reductions and scans). As part of enabling this, the _copy_type_metadata method was refactored to have different column types override it to perform column-specific tasks (necessary because DecimalColumn has some special handling), which should also provide some marginal performance improvements for methods that copy metadata.

This PR (and some to follow) are designed to help simplify the ongoing [Array refactoring](#8273).

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Keith Kraus (https://github.com/kkraus14)
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/8278
