+++ python/cudf/cudf/core/column/categorical.py
@@ -1506,6 +1506,28 @@ class CategoricalColumn(column.ColumnBase):
             offset=codes_col.offset,
         )
 
+    def _copy_type_metadata(
+        self: CategoricalColumn, other: ColumnBase
+    ) -> ColumnBase:
+        """Copies type metadata from self onto other, returning a new column.
+
+        In addition to the default behavior, if `other` is not a
+        CategoricalColumn, we assume other is a column of codes, and return a
+        CategoricalColumn composed of `other`  and the categories of `self`.
+        """
+        if not isinstance(other, cudf.core.column.CategoricalColumn):
+            other = column.build_categorical_column(
+                categories=self.categories,
+                codes=column.as_column(other.base_data, dtype=other.dtype),
+                mask=other.base_mask,
+                ordered=self.ordered,
+                size=other.size,
+                offset=other.offset,
+                null_count=other.null_count,
+            )
+        # Have to ignore typing here because it misdiagnoses super().
+        return super()._copy_type_metadata(other)  # type: ignore
+
 
 def _create_empty_categorical_column(
     categorical_column: CategoricalColumn, dtype: "CategoricalDtype"
