+++ python/cudf/cudf/core/column/numerical.py
@@ -2,20 +2,16 @@
 
 from __future__ import annotations
 
-import builtins
-from numbers import Number
 from types import SimpleNamespace
-from typing import Any, Callable, Mapping, Sequence, Tuple, Union, cast
+from typing import Any, Mapping, Sequence, Tuple, Union, cast
 
 import cupy
 import numpy as np
 import pandas as pd
-from numba import cuda, njit
 from pandas.api.types import is_integer_dtype
 
 import cudf
 from cudf import _lib as libcudf
-from cudf._lib.quantiles import quantile as cpp_quantile
 from cudf._typing import BinaryOperand, ColumnLike, Dtype, DtypeObj, ScalarLike
 from cudf.core.buffer import Buffer
 from cudf.core.column import (
@@ -23,7 +19,6 @@ from cudf.core.column import (
     as_column,
     build_column,
     column,
-    column_empty,
     string,
 )
 from cudf.core.dtypes import Decimal64Dtype
@@ -37,8 +32,10 @@ from cudf.utils.dtypes import (
     to_cudf_compatible_scalar,
 )
 
+from .numerical_base import NumericalBaseColumn
 
-class NumericalColumn(ColumnBase):
+
+class NumericalColumn(NumericalBaseColumn):
     def __init__(
         self,
         data: Buffer,
@@ -91,7 +88,7 @@ class NumericalColumn(ColumnBase):
         ).any()
 
     @property
-    def __cuda_array_interface__(self) -> Mapping[builtins.str, Any]:
+    def __cuda_array_interface__(self) -> Mapping[str, Any]:
         output = {
             "shape": (len(self),),
             "strides": (self.dtype.itemsize,),
@@ -167,9 +164,6 @@ class NumericalColumn(ColumnBase):
             out_dtype = "bool"
         lhs, rhs = (self, rhs) if not reflect else (rhs, self)
         return libcudf.binaryop.binaryop(lhs, rhs, binop, out_dtype)
-
-    def _apply_scan_op(self, op: str) -> ColumnBase:
-        return libcudf.reduce.scan(op, self, True)
 
     def normalize_binop_value(
         self, other: ScalarLike
@@ -264,43 +258,6 @@ class NumericalColumn(ColumnBase):
             return self
         return libcudf.unary.cast(self, dtype)
 
-    def reduce(self, op: str, skipna: bool = None, **kwargs) -> float:
-        min_count = kwargs.pop("min_count", 0)
-        preprocessed = self._process_for_reduction(
-            skipna=skipna, min_count=min_count
-        )
-        if isinstance(preprocessed, ColumnBase):
-            return libcudf.reduce.reduce(op, preprocessed, **kwargs)
-        else:
-            return cast(float, preprocessed)
-
-    def sum(
-        self, skipna: bool = None, dtype: Dtype = None, min_count: int = 0
-    ) -> float:
-        return self.reduce(
-            "sum", skipna=skipna, dtype=dtype, min_count=min_count
-        )
-
-    def product(
-        self, skipna: bool = None, dtype: Dtype = None, min_count: int = 0
-    ) -> float:
-        return self.reduce(
-            "product", skipna=skipna, dtype=dtype, min_count=min_count
-        )
-
-    def mean(self, skipna: bool = None, dtype: Dtype = np.float64) -> float:
-        return self.reduce("mean", skipna=skipna, dtype=dtype)
-
-    def var(
-        self, skipna: bool = None, ddof: int = 1, dtype: Dtype = np.float64
-    ) -> float:
-        return self.reduce("var", skipna=skipna, dtype=dtype, ddof=ddof)
-
-    def std(
-        self, skipna: bool = None, ddof: int = 1, dtype: Dtype = np.float64
-    ) -> float:
-        return self.reduce("std", skipna=skipna, dtype=dtype, ddof=ddof)
-
     def _process_values_for_isin(
         self, values: Sequence
     ) -> Tuple[ColumnBase, ColumnBase]:
@@ -316,163 +273,6 @@ class NumericalColumn(ColumnBase):
             rhs = rhs.astype(lhs.dtype)
 
         return lhs, rhs
-
-    def sum_of_squares(self, dtype: Dtype = None) -> float:
-        return libcudf.reduce.reduce("sum_of_squares", self, dtype=dtype)
-
-    def kurtosis(self, skipna: bool = None) -> float:
-        skipna = True if skipna is None else skipna
-
-        if len(self) == 0 or (not skipna and self.has_nulls):
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-
-        self = self.nans_to_nulls().dropna()  # type: ignore
-
-        if len(self) < 4:
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-
-        n = len(self)
-        miu = self.mean()
-        m4_numerator = ((self - miu) ** self.normalize_binop_value(4)).sum()
-        V = self.var()
-
-        if V == 0:
-            return 0
-
-        term_one_section_one = (n * (n + 1)) / ((n - 1) * (n - 2) * (n - 3))
-        term_one_section_two = m4_numerator / (V ** 2)
-        term_two = ((n - 1) ** 2) / ((n - 2) * (n - 3))
-        kurt = term_one_section_one * term_one_section_two - 3 * term_two
-        return kurt
-
-    def skew(self, skipna: bool = None) -> ScalarLike:
-        skipna = True if skipna is None else skipna
-
-        if len(self) == 0 or (not skipna and self.has_nulls):
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-
-        self = self.nans_to_nulls().dropna()  # type: ignore
-
-        if len(self) < 3:
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-
-        n = len(self)
-        miu = self.mean()
-        m3 = (((self - miu) ** self.normalize_binop_value(3)).sum()) / n
-        m2 = self.var(ddof=0)
-
-        if m2 == 0:
-            return 0
-
-        unbiased_coef = ((n * (n - 1)) ** 0.5) / (n - 2)
-        skew = unbiased_coef * m3 / (m2 ** (3 / 2))
-        return skew
-
-    def quantile(
-        self, q: Union[float, Sequence[float]], interpolation: str, exact: bool
-    ) -> NumericalColumn:
-        if isinstance(q, Number) or cudf.utils.dtypes.is_list_like(q):
-            np_array_q = np.asarray(q)
-            if np.logical_or(np_array_q < 0, np_array_q > 1).any():
-                raise ValueError(
-                    "percentiles should all be in the interval [0, 1]"
-                )
-        # Beyond this point, q either being scalar or list-like
-        # will only have values in range [0, 1]
-        result = self._numeric_quantile(q, interpolation, exact)
-        if isinstance(q, Number):
-            return (
-                cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-                if result[0] is cudf.NA
-                else result[0]
-            )
-        return result
-
-    def median(self, skipna: bool = None) -> NumericalColumn:
-        skipna = True if skipna is None else skipna
-
-        if not skipna and self.has_nulls:
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-
-        # enforce linear in case the default ever changes
-        return self.quantile(0.5, interpolation="linear", exact=True)
-
-    def _numeric_quantile(
-        self, q: Union[float, Sequence[float]], interpolation: str, exact: bool
-    ) -> NumericalColumn:
-        quant = [float(q)] if not isinstance(q, (Sequence, np.ndarray)) else q
-        # get sorted indices and exclude nulls
-        sorted_indices = self.as_frame()._get_sorted_inds(
-            ascending=True, na_position="first"
-        )
-        sorted_indices = sorted_indices[self.null_count :]
-
-        return cpp_quantile(self, quant, interpolation, sorted_indices, exact)
-
-    def cov(self, other: ColumnBase) -> float:
-        if (
-            len(self) == 0
-            or len(other) == 0
-            or (len(self) == 1 and len(other) == 1)
-        ):
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-
-        result = (self - self.mean()) * (other - other.mean())
-        cov_sample = result.sum() / (len(self) - 1)
-        return cov_sample
-
-    def corr(self, other: ColumnBase) -> float:
-        if len(self) == 0 or len(other) == 0:
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-
-        cov = self.cov(other)
-        lhs_std, rhs_std = self.std(), other.std()
-
-        if not cov or lhs_std == 0 or rhs_std == 0:
-            return cudf.utils.dtypes._get_nan_for_dtype(self.dtype)
-        return cov / lhs_std / rhs_std
-
-    def round(self, decimals: int = 0) -> NumericalColumn:
-        """Round the values in the Column to the given number of decimals.
-        """
-        return libcudf.round.round(self, decimal_places=decimals)
-
-    def applymap(
-        self, udf: Callable[[ScalarLike], ScalarLike], out_dtype: Dtype = None
-    ) -> ColumnBase:
-        """Apply an elementwise function to transform the values in the Column.
-
-        Parameters
-        ----------
-        udf : function
-            Wrapped by numba jit for call on the GPU as a device function.
-        out_dtype  : numpy.dtype; optional
-            The dtype for use in the output.
-            By default, use the same dtype as *self.dtype*.
-
-        Returns
-        -------
-        result : Column
-            The mask is preserved.
-        """
-        if out_dtype is None:
-            out_dtype = self.dtype
-
-        core = njit(udf)
-
-        # For non-masked columns
-        @cuda.jit
-        def kernel_applymap(values, results):
-            i = cuda.grid(1)
-            # in range?
-            if i < values.size:
-                # call udf
-                results[i] = core(values[i])
-
-        results = column_empty(self.size, dtype=out_dtype)
-        values = self.data_array_view
-        kernel_applymap.forall(self.size)(values, results)
-        return as_column(results)
 
     def default_na_value(self) -> ScalarLike:
         """Returns the default NA value for this column
